config objects and support for the serial button

+++ events/button_pressed.sh
@@ -6,9 +6,11 @@ if [ -f $DIR/env ] ; then
   source "$DIR/env"
 fi
 
-killall mplayer
-blink1-tool --rgb FF9900 -l1
-mplayer -playlist http://streaming.radio.co/s774887f7b/listen.m3u -really-quiet > /dev/null 2> /dev/null < /dev/null &
+if [ -n "${RADIO-}" ] ; then
+  killall mplayer
+  blink1-tool --rgb FF9900 -l1
+  mplayer -playlist http://streaming.radio.co/s774887f7b/listen.m3u -really-quiet > /dev/null 2> /dev/null < /dev/null &
+fi
 
 MSG="It's coffee time. $($DIR/consuloftheday/query.sh)"
 echo $MSG
